// frontend/Lastnext/app/lib/types.tsx
export interface User {
  id: string;
  username: string;
  email: string | null;
  profile_image: string | null;
  positions: string;
  properties: Property[];
  accessToken: string;
  refreshToken: string;
  accessTokenExpires?: number;
  sessionToken?: string;
  users?: number[];
  created_at: string;
}

export interface ProfileImage {
  profile_image: string | null;
  properties: Property[];
}
// ...
export type JobStatus = 'pending' | 'in_progress' | 'completed' | 'cancelled' | 'waiting_sparepart';
export type JobPriority = 'low' | 'medium' | 'high';

export interface JobImage {
  id: number;
  image_url: string;
  uploaded_by?: number | string | User | null;
  uploaded_at: string;
}

export interface Topic {
  id: number;
  title: string;
  description: string | null;
}

export interface TopicFromAPI {
  id: number;
  title: string;
  description: string | null;
}

export interface Room {
  room_id: number;
  name: string;
  room_type: string;
  is_active: boolean;
  created_at: string;
  properties?: (number | string | Property)[];
  property_id?: string | number | null;
}

export interface Job {
  id: number;
  job_id: string;
  description: string;
  status: JobStatus;
  priority: JobPriority;
  created_at: string;
  updated_at: string;
  completed_at: string | null;
  user: number | string | User;
  updated_by?: number | string | User | null;
  profile_image?: ProfileImage | null;
  images?: JobImage[];
  topics?: Topic[];
  rooms?: Room[];
  property_id?: string | number;
  properties?: Array<string | number | { property_id?: string | number; id?: string | number }>; // Added
  remarks?: string | undefined | null;
  is_defective?: boolean;
  image_urls?: string[];
  is_preventivemaintenance?: boolean;
}

export interface Property {
  property_id: string;
  name: string;
  description: string | null;
  created_at: string;
  rooms?: Room[];
  id: string;
  
}